requirements.txt
Django==3.1.5
djangorestframework==3.12.2
PyJWT==2.0.1
pytest==6.1.2
Pillow==8.1.0
